+++ analysis_options.yaml
@@ -10,7 +10,7 @@ include: package:lint/analysis_options.yaml
 # You might want to exclude auto-generated files from dart analysis
 analyzer:
   exclude:
-    #- '**.freezed.dart'
+  #- '**.freezed.dart'
 
 # You can customize the lint rules set to your own liking. A list of all rules
 # can be found at https://dart-lang.github.io/linter/lints/options/options.html
@@ -22,8 +22,8 @@ linter:
     # The new tabs vs. spaces. Choose wisely
     prefer_single_quotes: true
 
-    # Good packages document everything
-    public_member_api_docs: true
+    #    # Good packages document everything
+    #    public_member_api_docs: true
 
     # Blindly follow the Flutter code style, which prefers types everywhere
     always_specify_types: true
