+++ lib/admiralui_flutter.dart
@@ -1,8 +1,3 @@
-/// Flutter package for [Eva Icons](https://akveo.github.io/eva-icons/). Eva Icons is a pack of more than 480 beautifully crafted Open Source icons for common actions and items.
-///
-/// This package brings [Eva Icons](https://akveo.github.io/eva-icons/) pack available as Flutter Icons. It has beautiful outlined and rounded Icons.
-///
-/// Blog Post: [Beautiful Outlined and Rounded Icons in Flutter with Eva Icons](https://medium.com/@piyushmaurya23/beautiful-outlined-and-rounded-icons-in-flutter-with-eva-icons-e9bf86a7c4e9)
 library admiralui_flutter;
 
 export 'src/admiral_icons_flutter.dart';
