+++ lib/gen/fonts.gen.dart
@@ -1,0 +1,21 @@
+/// GENERATED CODE - DO NOT MODIFY BY HAND
+/// *****************************************************
+///  FlutterGen
+/// *****************************************************
+
+// coverage:ignore-file
+// ignore_for_file: type=lint
+// ignore_for_file: directives_ordering,unnecessary_import
+
+class FontFamily {
+  FontFamily._();
+
+  /// Font family: AdmiralIcons
+  static const String admiralIcons = 'AdmiralIcons';
+
+  /// Font family: sfProDisplay
+  static const String sfprodisplay = 'sfProDisplay';
+
+  /// Font family: sfProText
+  static const String sfprotext = 'sfProText';
+}
