--- lib/src/Widgets/Controlls/icon_direction.dart
@@ -1,6 +1,0 @@
-import 'package:flutter/material.dart';
-
-enum IconDirection {
-  left,
-  right,
-}
