+++ lib/src/widgets/controls/control_parameter.dart
@@ -1,0 +1,28 @@
+enum ControlState {
+  normal,
+  highlighted,
+  disabled,
+  selected,
+}
+
+class ControlParameter<T> {
+  final Map<String, T> _parameters = {};
+
+  void set(T parameter, ControlState forState) {
+    _parameters[forState.name] = parameter;
+  }
+
+  T? parameter(ControlState forState) {
+    return _parameters[forState.name];
+  }
+
+  T unsafeParameter(ControlState forState) {
+    final result = _parameters[forState.name];
+    if (result != null) {
+      return result;
+    } else {
+      throw "Parameter with the name '${forState.name}' is undeclared in the "
+          'ControlState.';
+    }
+  }
+}
