+++ lib/src/widgets/controls/icon_direction.dart
@@ -1,0 +1,4 @@
+enum IconDirection {
+  left,
+  right,
+}
